+++ dsipts/src/dsipts/data_structure/data_structure.py
@@ -451,7 +451,8 @@ class TimeSeries():
 
         idx_target = []
         for c in self.target_variables:
-            idx_target.append(self.past_variables.index(c))
+            if c in self.past_variables:
+                idx_target.append(self.past_variables.index(c))
             
         idx_target_future = []
         
@@ -828,7 +829,7 @@ class TimeSeries():
         self.model.to(torch.device("cuda:0" if torch.cuda.is_available() else "cpu"))
         if self.model.can_be_compiled():
             try:
-                self.model = torch.compile(self.model)
+                #self.model = torch.compile(self.model)
                 beauty_string('Model COMPILED','block',self.verbose)
 
             except: